refactor: resolvendo problemas da pull request
